+++ src/connect/components/drive-view/index.ts
@@ -1,2 +1,1 @@
 export * from './drive-view';
-export * from './utils';
