--- src/connect/components/drive-view/utils.ts
@@ -1,8 +1,0 @@
-import {
-    CLOUD_DRIVE,
-    DriveTreeItem,
-    LOCAL_DRIVE,
-    PUBLIC_DRIVE,
-    TreeItem,
-    driveTypes,
-} from '@/connect';
