--- src/connect/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './tree-view';
-export * from './useENSInfo';
