--- src/connect/hooks/tree-view/index.ts
@@ -1,8 +1,0 @@
-export * from './useFilerPathContent';
-export * from './useGetDriveParent';
-export * from './useGetItemById';
-export * from './useGetItemByPath';
-export * from './useGetPathContent';
-export * from './useItemActions';
-export * from './usePathContent';
-export * from './utils';
