--- src/connect/hooks/tree-view/useFilerPathContent/index.ts
@@ -1,1 +1,0 @@
-export * from './useFilterPathContent';
