--- src/connect/hooks/tree-view/useGetDriveParent/index.ts
@@ -1,1 +1,0 @@
-export * from './useGetDriveParent';
