--- src/connect/hooks/tree-view/useGetItemById/index.ts
@@ -1,1 +1,0 @@
-export * from './useGetItemById';
