--- src/connect/hooks/tree-view/useGetItemByPath/index.ts
@@ -1,1 +1,0 @@
-export * from './useGetItemByPath';
