--- src/connect/hooks/tree-view/useGetItemByPath/useGetItemByPath.ts
@@ -1,11 +1,0 @@
-import { useItemsContext } from '@/connect';
-
-export const useGetItemByPath = () => {
-    const { items } = useItemsContext();
-
-    const getItemByPath = (path: string) => {
-        return items.find(item => item.path === path);
-    };
-
-    return getItemByPath;
-};
