--- src/connect/hooks/tree-view/useGetPathContent/index.ts
@@ -1,1 +1,0 @@
-export * from './useGetPathContent';
