--- src/connect/hooks/tree-view/useItemActions/index.ts
@@ -1,1 +1,0 @@
-export * from './useItemActions';
