--- src/connect/hooks/tree-view/usePathContent/index.ts
@@ -1,1 +1,0 @@
-export * from './usePathContent';
