--- src/connect/hooks/tree-view/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './getPathContent';
