+++ src/connect/index.ts
@@ -1,6 +1,5 @@
 export * from './components';
 export * from './constants';
 export * from './context';
-export * from './hooks';
 export * from './types';
 export * from './utils';
