+++ src/connect/utils/mocks/ui-drive-node.ts
@@ -1,5 +1,10 @@
-import { FILE, FOLDER, LOCAL } from '@/connect/constants';
-import { DocumentDriveDocument, FileNode, FolderNode } from '@/connect/context';
+import { CLOUD, FILE, FOLDER, LOCAL, PUBLIC } from '@/connect/constants';
+import {
+    DocumentDriveDocument,
+    FileNode,
+    FolderNode,
+    makeDriveNode,
+} from '@/connect/context';
 
 const mockDriveId = 'mock-drive-id';
 
@@ -107,3 +112,34 @@ export function makeMockDriveDocument(state?: {
 
     return mockDocumentDriveDocument;
 }
+export const mockLocalDrive = makeDriveNode(
+    makeMockDriveDocument({
+        global: {
+            id: 'local-drive',
+            name: 'Local drive',
+        },
+        local: { sharingType: LOCAL, availableOffline: false },
+    }),
+);
+
+export const mockPublicDrive = makeDriveNode(
+    makeMockDriveDocument({
+        global: {
+            id: 'public-drive',
+            name: 'Public drive',
+        },
+        local: { sharingType: PUBLIC, availableOffline: true },
+    }),
+);
+
+export const mockCloudDrive = makeDriveNode(
+    makeMockDriveDocument({
+        global: {
+            id: 'cloud-drive',
+            name: 'Cloud drive',
+        },
+        local: { sharingType: CLOUD, availableOffline: true },
+    }),
+);
+
+export const mockDriveNodes = [mockLocalDrive, mockPublicDrive, mockCloudDrive];
